% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_calc.R
\name{dist_calc}
\alias{dist_calc}
\title{Distance Calculation for Graph Nodes}
\usage{
dist_calc(g, formula = "Haversine")
}
\arguments{
\item{g}{An igraph object, with nodes that have 'latitude' and 'longitude' attributes}

\item{formula}{The distance calculation formula to use, either 'Haversine' or 'Euclidean'.}
}
\value{
A numeric vector with distances calculated between each pair of nodes in the graph.
}
\description{
This function calculates distances between each pair of nodes in the graph.
It supports both Haversine and Euclidean formula. The function requires a graph object
with nodes containing latitude and longitude attributes.
}
\examples{
# Assuming 'g' is a graph object with appropriate latitude and longitude attributes for each node
dist_calc(g, formula = 'Haversine')
dist_calc(g, formula = 'Euclidean')

}
