% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Calculate Euclidean Distance Between Two Points}
\usage{
euclidean(lon1, lat1, lon2, lat2)
}
\arguments{
\item{lon1}{Longitude of the first point in decimal degrees.}

\item{lat1}{Latitude of the first point in decimal degrees.}

\item{lon2}{Longitude of the second point in decimal degrees.}

\item{lat2}{Latitude of the second point in decimal degrees.}
}
\value{
The Euclidean distance between the two points.
}
\description{
This function calculates the Euclidean distance between two points. The Euclidean is the 'straight line' distance
between two points in a two-dimensional space. This function takes the coordinates of two points (longitude and latitude)
and calculates the straight distance between them, assuming flat Earth approximation.
}
\examples{
euclidean(-73.9851, 40.7580, -0.1278, 51.5074)
# Euclidean distance between New York City and London.

}
