% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSDI-package.R
\docType{package}
\name{rSDI-package}
\alias{rSDI}
\alias{rSDI-package}
\title{rSDI: Spatial Dispersion Index (SDI) Family of Metrics for Spatial/Geographic Networks}
\description{
Spatial Dispersion Index (SDI) is a generalized measurement index, or rather a family of indices to evaluate spatial dispersion of movements/flows in a network in a problem neutral way as described in: Gencer (2023) \doi{10.1007/s12061-023-09545-8}. This package computes and optionally visualizes this index with minimal hassle.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ehengirmen/rSDI}
  \item Report bugs at \url{https://github.com/ehengirmen/rSDI/issues}
}

}
\author{
\strong{Maintainer}: Mehmet Gençer \email{mehmetgencer@yahoo.com} (\href{https://orcid.org/0000-0003-1717-8668}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Mücahit Zor \email{mucahit.zor01@gmail.com}
  \item Engin Hengirmen \email{ehengirmen@gmail.com}
}

}
\keyword{internal}
