% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TurkiyeMigration.flows}
\alias{TurkiyeMigration.flows}
\title{Türkiye migration network flows data}
\format{
## `TurkiyeMigration.flows`
A data frame with 6480 rows and 3 columns:
\describe{
  \item{from, to}{codes of origin and arrival province}
  \item{weight}{number of people migrated}
}
}
\source{
<https://data.tuik.gov.tr/Kategori/GetKategori?p=Nufus-ve-Demografi-109>
}
\usage{
TurkiyeMigration.flows
}
\description{
This data frame contains the data on migration of people between Türkiye's provinces
in the period 2016-2017-2018.
This is a consolidated version of raw data from Turkish Statistical Institute.
}
\details{
Each row contains the number of people migrated in the "weight" column. 'from' and 'to'
columns include the codes of provinces as used in the Türkiye statistical coding system.

The familiar names of provinces and their locations are to be found in a separate data frame
named TurkiyeMigration.nodes
}
\keyword{datasets}
