% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TurkiyeMigration.nodes}
\alias{TurkiyeMigration.nodes}
\title{Türkiye migration network provinces/nodes data}
\format{
## `TurkiyeMigration.nodes`
A data frame with 6480 rows and 3 columns:
\describe{
  \item{id}{codes of province as used in the TurkiyeMigration.flows data frame}
  \item{label}{Name of the province capital city}
  \item{longitude, latitude}{spatial coordinates of the province capital}
}
}
\source{
The latitude/longitude of Turkish province capitals was shared as a courtesy of Başarsoft LLC
}
\usage{
TurkiyeMigration.nodes
}
\description{
This data frame contains the data on Türkiye's provinces as used in the migration flows
data frame (TurkiyeMigration.flow)
}
\details{
Each row contains the code of the province as used in TurkiyeMigration.flows data frame,
it is known name/label, and latitude/longitude of the province central.
}
\keyword{datasets}
