% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDIvariants.R
\name{unweightedAllVerticesSDI}
\alias{unweightedAllVerticesSDI}
\title{Not for explicit use.}
\usage{
unweightedAllVerticesSDI(g, mode = "all")
}
\arguments{
\item{mode}{}
}
\description{
Not for explicit use.
}
