% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDIvariants.R
\name{unweightedNetworkSDI}
\alias{unweightedNetworkSDI}
\title{Not for explicit use.}
\usage{
unweightedNetworkSDI(g)
}
\arguments{
\item{g}{the graph}
}
\value{
a numerical SDI value
}
\description{
Not for explicit use.
}
