% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDIvariants.R
\name{unweightedSingleVertexSDI}
\alias{unweightedSingleVertexSDI}
\title{Not for explicit use.}
\usage{
unweightedSingleVertexSDI(g, v, mode = all)
}
\arguments{
\item{g}{the graph}

\item{v}{the vertex}

\item{mode}{directionality 'undirected', 'in', 'out', or 'all'}
}
\value{
a numerical SDI value
}
\description{
Not for explicit use.
}
