% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDI.R
\name{variantParser}
\alias{variantParser}
\title{variantParser for SDI variant short-codes.
This is a helper function and not intended for explicit use.}
\usage{
variantParser(variant)
}
\arguments{
\item{variant}{a three letter short code for level, weight, and direction of SDI calculation.}
}
\value{
A list of explicit level, weight.use and directionality parameters
}
\description{
variantParser for SDI variant short-codes.
This is a helper function and not intended for explicit use.
}
\examples{
variantParser("vuw")
}
