% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDIvariants.R
\name{weightedAllVerticesSDI}
\alias{weightedAllVerticesSDI}
\title{Not for explicit use.}
\usage{
weightedAllVerticesSDI(g, mode = "all")
}
\arguments{
\item{g}{the graph}

\item{mode}{directionality 'undirected', 'in', 'out', or 'all'}
}
\value{
a vector of values
}
\description{
Not for explicit use.
}
