% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDIvariants.R
\name{weightedNetworkSDI}
\alias{weightedNetworkSDI}
\title{Not for explicit use.}
\usage{
weightedNetworkSDI(g)
}
\arguments{
\item{g}{the graph}
}
\value{
a numerical SDI value
}
\description{
Not for explicit use.
}
