% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDIvariants.R
\name{weightedSingleVertexSDI}
\alias{weightedSingleVertexSDI}
\title{Not for explicit use.}
\usage{
weightedSingleVertexSDI(g, v, mode = all)
}
\arguments{
\item{mode}{}
}
\description{
Not for explicit use.
}
