% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_calc.R
\name{euclidean}
\alias{euclidean}
\title{Calculate Euclidean Distance Between Two Points}
\usage{
euclidean(x1, y1, x2, y2)
}
\arguments{
\item{x1}{X-coordinate of the first point.}

\item{y1}{Y-coordinate of the first point.}

\item{x2}{X-coordinate of the second point.}

\item{y2}{Y-coordinate of the second point.}
}
\value{
The Euclidean distance between the two points.
}
\description{
This function calculates the Euclidean distance between two points. The Euclidean is the 'straight line' distance
between two points in a two-dimensional space. This function takes the coordinates of two points (longitude and latitude)
and calculates the straight distance between them, assuming flat Earth approximation.
}
\examples{
euclidean(1, 2, 4, 6)
# Euclidean distance between points (1, 2) and (4, 6).

}
