% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDI.R
\name{SDIcomputer}
\alias{SDIcomputer}
\title{SDIcomputer() is a helper function to compute given SDI variant for the given graph object.
Not intended for explicit use. Called automatically by the `SDI()` function.}
\usage{
SDIcomputer(g, level, weight.use, directionality, mode)
}
\arguments{
\item{mode}{}
}
\description{
SDIcomputer() is a helper function to compute given SDI variant for the given graph object.
Not intended for explicit use. Called automatically by the `SDI()` function.
}
\examples{
TMgraph <- graph_from_data_frame(TurkiyeMigration.flows, directed=TRUE, TurkiyeMigration.nodes)
SDIcomputer(TMgraph,"vertex","weighted","in")
}
